+++ grammargen/z_diag_scala_root_test.go
@@ -320,7 +320,7 @@ func TestDiagScalaRootRuntime(t *testing.T) {
 					state, state+1, lexState, lexProbeByte, tok.Symbol, tokName, tok.EndByte-tok.StartByte)
 				if state < len(diagStateToMode) && diagStateToMode[state] < len(diagLexModeOffsets) {
 					modeIdx := diagStateToMode[state]
-					computedLexState := uint16(diagLexModeOffsets[modeIdx])
+					computedLexState := uint32(diagLexModeOffsets[modeIdx])
 					diagLexer := gotreesitter.NewLexer(diagLexStates, src[lexProbeByte:])
 					diagTok := diagLexer.Next(computedLexState)
 					diagName := ""
